creation of dedicated fixtures
